--- src/main/java/com/example/loveislandsimulator/HomeController.java
@@ -1,14 +1,0 @@
-package com.example.loveislandsimulator;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HomeController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-}
