--- src/main/java/com/example/loveislandsimulator/IslanderBehaviorStrategy.java
@@ -1,7 +1,0 @@
-package com.example.loveislandsimulator;
-
-public interface IslanderBehaviorStrategy {
-    void perform(Islander islander, ChallengeCommand challenge);
-}
-
-
