+++ src/main/java/com/loveislandsimulator/challenges/PhysicalChallenge.java
@@ -1,4 +1,7 @@
-package com.example.loveislandsimulator;
+package com.loveislandsimulator.challenges;
+
+import com.loveislandsimulator.models.ChallengeCommand;
+import com.loveislandsimulator.models.Islander;
 
 public class PhysicalChallenge implements ChallengeCommand {
     @Override
