+++ src/main/java/com/loveislandsimulator/controllers/HomeController.java
@@ -1,0 +1,10 @@
+package com.loveislandsimulator.controllers;
+
+import javafx.fxml.FXML;
+
+public class HomeController {
+    @FXML
+    protected void onHelloButtonClick() {
+
+    }
+}
