+++ src/main/java/com/loveislandsimulator/models/ChallengeCommand.java
@@ -1,4 +1,4 @@
-package com.example.loveislandsimulator;
+package com.loveislandsimulator.models;
 
 public interface ChallengeCommand {
     void execute(Islander islander);
