+++ src/main/java/com/loveislandsimulator/models/Islander.java
@@ -1,4 +1,6 @@
-package com.example.loveislandsimulator;
+package com.loveislandsimulator.models;
+
+import com.loveislandsimulator.strategies.IslanderBehaviorStrategy;
 
 import java.io.Serializable;
